Tried doing live georef in python, didnt work, but will keep it for future reference. And started including new networking messages for georef data

+++ Sonarpublish/src/common/sonarmeta.h
@@ -84,7 +84,7 @@ typedef struct
 #define GST_SONAR_TELEMETRY_PRESENCE_N_FIELDS  (7)
 #define GST_SONAR_TELEMETRY_PRESENCE_FULL      ((1 << GST_SONAR_TELEMETRY_PRESENCE_N_FIELDS) - 1)
 
-typedef float GstSonarTelemetryField;
+typedef double GstSonarTelemetryField;
 
 // contains telemetry data. which data is specified in the presence field
 typedef struct
